refactor: Move manager classes to managers folder
