+++ source/gallery.ts
@@ -1,73 +1,7 @@
 import type { Hitomi } from './hitomi';
 import { Image, Video } from './media';
 import { Language, Tag } from './tag';
-import { RESOURCE_DOMAIN, DEDICATED_TAG_PROPERTIES } from './internal/constants';
-import { defineProperties } from './internal/functions';
-import { Base, IndexProvider } from './internal/structures';
-import { HitomiError } from './error';
-import type { Node } from './internal/types';
-import { ResponseType } from '@platform';
-
-/**
- * Sorting options for listing galleries.
- *
- * @enum {string}
- */
-export const enum SortType {
-	DateAdded = 'added',
-	DatePublished = 'published',
-	Random = 'random',
-	PopularityDay = 'today',
-	PopularityWeek = 'week',
-	PopularityMonth = 'month',
-	PopularityYear = 'year'
-}
-
-/**
- * Pagination options for listing galleries.
- * 
- * @see {@link GalleryOptions}
- */
-export interface PageOptions {
-	/**
-	 * Zero-based page index.
-	 * 
-	 * @default 0
-	 */
-	index?: number;
-	/**
-	 * Number of galleries per page.
-	 * 
-	 * @default 25
-	 */
-	size?: number;
-}
-
-/**
- * Filter options for listing galleries.
- * 
- * @see {@link GalleryManager.list}
- */
-export interface GalleryOptions {
-	/**
-	 * {@link Tag} instances to filter.
-	 */
-	tags?: Tag[];
-	/**
-	 * Title keyword to search for.
-	 */
-	title?: string;
-	/**
-	 * Sort type to order by.
-	 *
-	 * @default SortType.DateAdded
-	 */
-	orderBy?: SortType;
-	/**
-	 * Pagination options.
-	 */
-	page?: PageOptions;
-}
+import { Base } from './internal/structures';
 
 /**
  * Title associated with a gallery.
@@ -275,414 +209,5 @@ export class Gallery extends TranslatedGallery {
 	 */
 	public getThumbnails(): [Image, Image] {
 		return [this['files'][0], this['files'][Math.floor(this['files']['length'] / 2)]]
-	}
-}
-
-/**
- * Manager for retrieving and listing {@link Gallery} instances.
- *
- * @see {@link Hitomi}
- */
-export class GalleryManager extends Base {
-	// @internal
-	private readonly index!: IndexProvider;
-
-	// @internal
-	constructor(hitomi: Hitomi) {
-		super(hitomi);
-
-		defineProperties(this, {
-			index: new IndexProvider(hitomi, 'galleries')
-		});
-	}
-
-	/**
-	 * Retrieves a {@link Gallery} by a unique identifier.
-	 *
-	 * @param {number} id Unique gallery identifier.
-	 * @returns {Promise<Gallery>} Promise that resolves to a {@link Gallery} instance.
-	 */
-	public async retrieve(id: number): Promise<Gallery> {
-		const rawGallery: {
-			id: string;
-			type: 'doujinshi' | 'manga' | 'artistcg' | 'gamecg' | 'anime';
-			galleryurl: string;
-			language: string | null;
-			language_localname: string | null;
-			title: string;
-			japanese_title: string | null;
-			artists: {
-				artist: string;
-			}[] | null;
-			groups: {
-				group: string;
-			}[] | null;
-			parodys: {
-				parody: string;
-			}[] | null;
-			characters: {
-				character: string;
-			}[] | null;
-			tags: {
-				tag: string;
-				male?: string | number;
-				female?: string | number;
-			}[];
-			files: {
-				hash: string;
-				name: string;
-				width: number;
-				height: number;
-				hasavif?: number;
-			}[];
-			languages: {
-				galleryid: number;
-				name: string;
-				language_localname: string;
-				url: string;
-			}[];
-			related: number[];
-			blocked: number;
-			date: string;
-			datepublished: string | null;
-			videofilename: string | null;
-		} = JSON.parse((await this['hitomi'].request(RESOURCE_DOMAIN, '/galleries/' + id + '.js', ResponseType['TEXT'])).slice(18));
-		const dedicatedTags: [Tag[], Tag[], Tag[], Tag[]] = [[] /* artists */, [] /* groups */, [] /* series */, [] /* characters */];
-		const tags: Tag[] = [];
-		const files: Image[] = [];
-		const translations: TranslatedGallery[] = [];
-
-		let i: number = 0;
-		let type: Tag['type'];
-
-		for(; i < DEDICATED_TAG_PROPERTIES['length']; i++) {
-			// @ts-expect-error - Typescript internal error
-			const dedicatedTagProperty: `${(typeof DEDICATED_TAG_PROPERTIES)[number]}s` = DEDICATED_TAG_PROPERTIES[i] + 's';
-
-			type = i !== 2 ? DEDICATED_TAG_PROPERTIES[i] as Tag['type'] : 'series';
-
-			if(rawGallery[dedicatedTagProperty]) {
-				for(let j: number = 0; j < rawGallery[dedicatedTagProperty]['length']; j++)
-					// @ts-expect-error - Typescript internal error
-					dedicatedTags[i].push(new Tag(this['hitomi'], type, rawGallery[dedicatedTagProperty][j][DEDICATED_TAG_PROPERTIES[i]]));
-			}
-		}
-
-		for(i = 0; i < rawGallery['tags']['length']; i++) {
-			if(Boolean(rawGallery['tags'][i]['male'])) {
-				type = 'male';
-			} else if(Boolean(rawGallery['tags'][i]['female'])) {
-				type = 'female';
-			} else {
-				type = 'tag';
-			}
-
-			tags.push(new Tag(this['hitomi'], type, rawGallery['tags'][i]['tag'], false));
-		}
-
-		const thumbnailIndex: number = Math.floor(rawGallery['files']['length'] / 2);
-
-		for(i = 0; i < rawGallery['files']['length']; i++) {
-			files.push(new Image(
-				this['hitomi'],
-				rawGallery['files'][i]['width'],
-				rawGallery['files'][i]['height'],
-				rawGallery['files'][i]['hash'],
-				rawGallery['files'][i]['name'],
-				Boolean(rawGallery['files'][i]['hasavif']),
-				true,
-				false,
-				!i || i === thumbnailIndex
-			));
-		}
-
-		for(i = 0; i < rawGallery['languages']['length']; i++) {
-			translations.push(new TranslatedGallery(
-				this['hitomi'],
-				rawGallery['languages'][i]['galleryid'],
-				new Language(
-					this['hitomi'],
-					rawGallery['languages'][i]['name'],
-					rawGallery['languages'][i]['language_localname']
-				),
-				rawGallery['languages'][i]['url']
-			));
-		}
-
-		const relations: GalleryReference[] = [];
-
-		for(i = 0; i < rawGallery['related']['length']; i++) {
-			relations.push(new GalleryReference(this['hitomi'], rawGallery['related'][i]));
-		}
-
-		return new Gallery(
-			this['hitomi'],
-			+rawGallery['id'],
-			rawGallery['language'] ? new Language(
-				this['hitomi'],
-				rawGallery['language'],
-				rawGallery['language_localname'] as Language['localName']
-			) : null,
-			// galleries with rearranged ids have strange urls
-			rawGallery['galleryurl'],
-			new Title(rawGallery['title'], rawGallery['japanese_title']),
-			rawGallery['type'],
-			dedicatedTags[0],
-			dedicatedTags[1],
-			dedicatedTags[2],
-			dedicatedTags[3],
-			tags,
-			files,
-			translations,
-			relations,
-			Boolean(rawGallery['blocked']),
-			new Date(rawGallery['date']),
-			rawGallery['datepublished'] ? new Date(rawGallery['datepublished']) : null,
-			rawGallery['videofilename'] ? new Video(this['hitomi'], files[1]['width'], files[1]['height'], rawGallery['videofilename'], files[1]['hash']) : null
-		);
-	}
-
-	// @internal
-	private async requestIds(url: [string, string], range?: string, isNegative: boolean = false): Promise<Set<Gallery['id']>> {
-		const view: DataView = await this['hitomi'].request(url[0], url[1], ResponseType['VIEW'], range);
-		const ids: Set<Gallery['id']> = new Set<Gallery['id']>();
-
-		for(let i: number = 0; i < view['byteLength']; i += 4) {
-			ids.add(view.getInt32(i));
-		}
-
-		if(isNegative) {
-			// negative flag
-			ids.add(0);
-		}
-
-		return ids;
-	}
-
-	// @internal
-	private static createNozomiUrl(options: {
-		tag?: Tag;
-		language?: string;
-		orderBy?: SortType;
-	} = {}): [string, string] {
-		const language: string = options['language'] || 'all';
-		let orderBy: string = '';
-
-		if(options['orderBy']) {
-			switch(options['orderBy']) {
-				case SortType['DatePublished']: {
-					orderBy = 'date/published';
-				}
-				case SortType['DateAdded']:
-				// shuffle array later
-				case SortType['Random']: {
-					break;
-				}
-
-				case SortType['PopularityDay']:
-				case SortType['PopularityWeek']:
-				case SortType['PopularityMonth']:
-				case SortType['PopularityYear']: {
-					orderBy = 'popular/' + options['orderBy'];
-
-					break;
-				}
-
-				default: {
-					// @ts-expect-error
-					throw HitomiError.OneOfState('OrderBy', SortType);
-				}
-			}
-		}
-
-		if(!options['tag'] || options['tag']['type'] === 'language') {
-			return [RESOURCE_DOMAIN, '/n/' + (orderBy || 'index') + '-' + language + '.nozomi'];
-		}
-
-		if(orderBy) {
-			orderBy += '/';
-		}
-
-		let area: string;
-
-		switch(options['tag']['type']) {
-			case 'male':
-			case 'female': {
-				area = 'tag/';
-				orderBy += options['tag']['type'] + ':';
-
-				break;
-			}
-
-			default: {
-				area = options['tag']['type'] + '/';
-			}
-		}
-
-		return [RESOURCE_DOMAIN, '/n/' + area + orderBy + encodeURIComponent(options['tag']['name']) + '-' + language + '.nozomi'];
-	}
-
-	// @internal
-	private createReferences(ids: Set<number>, shouldShuffle: boolean): GalleryReference[] {
-		const references: GalleryReference[] = [];
-
-		for(const id of ids) {
-			references.push(new GalleryReference(this['hitomi'], id));
-		}
-
-		if(shouldShuffle) {
-			let currentIndex: number = references['length'];
-			let targetIndex: number;
-
-			while(currentIndex) {
-				targetIndex = Math.floor(Math.random() * currentIndex--);
-
-				const temporary: GalleryReference = references[targetIndex];
-
-				references[targetIndex] = references[currentIndex];
-				references[currentIndex] = temporary;
-			}
-		}
-
-		return references;
-	}
-
-	/**
-	 * Lists {@link GalleryReference} entries that match the specified search criteria.
-	 *
-	 * When `options.page` is provided, only one non-language tag (optionally combined with a language tag) is allowed, and negative tags are not supported.
-	 *
-	 * When using `Popularity{Period}` in `options.orderBy`, the number of galleries may vary.
-	 *
-	 * @param {GalleryOptions} [options] Search options.
-	 * @returns {Promise<GalleryReference[]>} Promise that resolves to an array of {@link GalleryReference} instances.
-	 * @throws {HitomiError} Thrown when `page` is used with multiple tags or any negative tag.
-	 * @see {@link SortType}
-	 */
-	public async list(options: GalleryOptions = {}): Promise<GalleryReference[]> {
-		const idSets: Set<Gallery['id']>[] = [];
-		const isRandom: boolean = options['orderBy'] === 'random';
-		let language: string | undefined;
-		let i: number = 0;
-		let range: string | undefined;
-
-		if(options['page']) {
-			const size: number = options['page']['size'] ? options['page']['size'] * 4 : 100;
-			const start: number = options['page']['index'] ? options['page']['index'] * size : 0;
-
-			range = start + '-' + (start + size - 1);
-		}
-
-		if(options['tags'] && options['tags']['length']) {
-			// bring positive tags to front
-			const tags: Tag[] = options['tags'].slice().sort(function (a: Tag, b: Tag): number {
-				return (a['isNegative'] as unknown as number) - (b['isNegative'] as unknown as number);
-			});
-
-			if(tags[0]['isNegative']) {
-				i = -1;
-			} else {
-				for(; i < tags['length'] && !tags[i]['isNegative']; i++) {
-					if(tags[i]['type'] === 'language') {
-						language = tags[i]['name'];
-
-						break;
-					}
-				}
-
-				i = 0;
-			}
-
-			if(range) {
-				if(tags['length'] > 2 || tags['length'] === 2 && !language) {
-					throw new HitomiError('Page', 'used with multiple tags', false);
-				}
-
-				if(tags[tags['length'] - 1]['isNegative']) {
-					throw new HitomiError('Page', 'used with negative tag', false);
-				}
-
-				return this.createReferences(await this.requestIds(GalleryManager.createNozomiUrl({
-					tag: tags[+(tags[0]['type'] === 'language') /* selects non-language tag */],
-					orderBy: options['orderBy'],
-					language: language
-				}), range), isRandom);
-			}
-
-			idSets.push(await this.requestIds(GalleryManager.createNozomiUrl({
-				tag: tags[i++] /* if first tag is negative i becomes -1, therefore tags give undefined  */,
-				orderBy: options['orderBy'],
-				language: language
-			})));
-
-			for(; i < tags['length']; i++) {
-				if(tags[i]['type'] !== 'language' || !language && tags[i]['isNegative']) {
-					idSets.push(await this.requestIds(GalleryManager.createNozomiUrl({
-						tag: tags[i],
-						language: language
-					}), undefined, tags[i]['isNegative']));
-				}
-			}
-		} else {
-			const url: [string, string] = GalleryManager.createNozomiUrl({
-				orderBy: options['orderBy']
-			});
-
-			if(range) {
-				return this.createReferences(await this.requestIds(url, range), isRandom);
-			}
-
-			if(options['orderBy']) {
-				idSets.push(await this.requestIds(url));
-			}
-		}
-
-		if(options['title'] && options['title']['length']) {
-			const version: string = await this['index'].retrieve();
-			const title: string = options['title'].toLowerCase() + ' ';
-			const rootNode: Node | undefined = await this['index'].getNodeAtAddress(0n, version);
-
-			if(!rootNode) {
-				throw HitomiError['RootNodeEmpty'];
-			}
-
-			i /* currentIndex */ = 0;
-			let j /* nextIndex */: number = title.indexOf(' ');
-
-			while(j !== -1) {
-				if(j - i) {
-					const data: Node[1][number] | undefined = await this['index'].binarySearch(await this['hitomi'].hash(title.slice(i, j)), rootNode, version);
-
-					if(!data) {
-						return [];
-					}
-
-					idSets.push(await this.requestIds([RESOURCE_DOMAIN, '/galleriesindex/galleries.' + version + '.data'], (data[0] + 4n) + '-' + (data[0] + BigInt(data[1]) - 1n)));
-				}
-
-				i = j + 1;
-				j = title.indexOf(' ', i);
-			}
-		}
-
-		if(idSets['length']) {
-			for(i = 1; i < idSets['length']; i++) {
-				if(!idSets[0]['size']) {
-					return [];
-				}
-
-				const isNegative: boolean = idSets[i].has(0);
-
-				for(const id of idSets[0]) {
-					if(isNegative === idSets[i].has(id)) {
-						idSets[0].delete(id);
-					}
-				}
-			}
-		} else {
-			idSets.push(await this.requestIds(GalleryManager.createNozomiUrl()));
-		}
-
-		return this.createReferences(idSets[0], isRandom);
 	}
 }
