+++ source/hitomi.ts
@@ -1,6 +1,6 @@
 import type { ImageContext } from './internal/types';
-import { GalleryManager } from './gallery';
-import { TagManager } from './tag';
+import { GalleryManager } from './managers/gallery';
+import { TagManager } from './managers/tag';
 import { DEFAULT_HEADERS, RESOURCE_DOMAIN, MAXIMUM_AGE_PROPERTIES } from './internal/constants';
 import { defineProperties, capitalize } from './internal/functions';
 import { IndexProvider, Provider } from './internal/structures';
