+++ source/index.ts
@@ -1,9 +1,10 @@
 import { Hitomi } from './hitomi';
 
-export { Gallery, Title, TranslatedGallery, GalleryReference, GalleryManager, SortType, type GalleryOptions, type PageOptions } from './gallery';
+export * from './managers';
+export { Gallery, Title, TranslatedGallery, GalleryReference } from './gallery';
 export { Hitomi, type HitomiOptions } from './hitomi';
 export { Image, Video, Extension, ThumbnailSize } from './media';
-export { Language, Tag, TagManager, NameInitial } from './tag';
+export { Language, Tag } from './tag';
 export { HitomiError } from './error';
 export { ResponseType, type RequestContext } from '@platform';
