+++ source/managers/index.ts
@@ -1,0 +1,2 @@
+export * from './gallery';
+export * from './tag';
